services/ollama-service/requirements.txt
ollama==0.5.1
fastapi==0.116.0
uvicorn==0.35.0
requests==2.32.4
pydantic==2.11.7
PyYAML==6.0.1